Move the DisplayWork-publicationDate serialisation tests to the display lib
